dbms: Server: feature development. [#METR-15090]

+++ dbms/include/DB/IO/ReadBufferAIO.h
@@ -28,7 +28,7 @@ public:
 
 	void setMaxBytes(size_t max_bytes_read_);
 	off_t seek(off_t off, int whence = SEEK_SET);
-	size_t getPositionInFile() const noexcept { return pos_in_file - (working_buffer.end() - pos); }
+	off_t getPositionInFile() const noexcept { return pos_in_file - (working_buffer.end() - pos); }
 	std::string getFileName() const noexcept override { return filename; }
 	int getFD() const noexcept override { return fd; }
